refactor: move runnertest helper under internal/ to firewall singleton swap
Relocates pkg/runner/runnertest to internal/runnertest so the Go
compiler blocks imports from outside this module. The singleton-swap
helper still exists for cross-package tests (e.g., shell handler) but
external modules can no longer pull it in, addressing the production
API-surface concern raised on auth_manager_testing.go.
